fpsos-bot/requirements.txt
# Discord Bot Dependencies
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Optional: AI Analysis
openai==1.6.1

# Optional: Database
aiosqlite==0.19.0

# Firecrawl for Web Search
firecrawl-py>=0.0.18
